Fix bug when displaying the complex events + collapse identical sequences
The events were off by 1bp. This new version uses 1-based coordinates for the $min_from variable
to simplify things and avoid further mistakes.

Also, the script now collapses all the events that lead to the same sequence. In some cases, reads
with exactly the same sequence can be aligned differently by bowtie2 if their quality scores are
different. This has been observed in a case where the WT 'TGAACCAG' was mutated into 'TGAAAACAG',
sometimes aligned as 'TG(A)AAaCAG' and sometimes as 'TGAA(A)aCAG'.